#14. Renamed Devices.Simulator to ScenarioSimulator
