+++ src/Implementation/Simulator/ScenarioSimulator.Tests/Properties/AssemblyInfo.cs
@@ -7,11 +7,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Microsoft.Practices.IoTJourney.Devices.Simulator.Tests")]
+[assembly: AssemblyTitle("Microsoft.Practices.IoTJourney.ScenarioSimulator.Tests")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Microsoft.Practices.IoTJourney.Devices.Simulator.Tests")]
+[assembly: AssemblyProduct("Microsoft.Practices.IoTJourney.ScenarioSimulator.Tests")]
 [assembly: AssemblyCopyright("Copyright ©  2015")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
