+++ src/Implementation/Simulator/ScenarioSimulator.Tests/WhenSendingEvents.cs
@@ -4,7 +4,7 @@
 using Microsoft.Practices.IoTJourney.Devices.Events;
 using Xunit;
 
-namespace Microsoft.Practices.IoTJourney.Devices.Simulator.Tests
+namespace Microsoft.Practices.IoTJourney.ScenarioSimulator.Tests
 {
     public class WhenSendingEvents
     {
