+++ src/Implementation/Simulator/ScenarioSimulator/EventSender.cs
@@ -5,11 +5,11 @@ using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
 using Microsoft.Practices.IoTJourney.Devices.Events;
-using Microsoft.Practices.IoTJourney.Devices.Simulator.Instrumentation;
 using Microsoft.Practices.IoTJourney.Logging;
+using Microsoft.Practices.IoTJourney.ScenarioSimulator.Instrumentation;
 using Microsoft.ServiceBus.Messaging;
 
-namespace Microsoft.Practices.IoTJourney.Devices.Simulator
+namespace Microsoft.Practices.IoTJourney.ScenarioSimulator
 {
     public class EventSender : IEventSender
     {
