+++ src/RunFromConsole/ScenarioSimulator.ConsoleHost/App.config
