+++ src/RunFromConsole/ScenarioSimulator.ConsoleHost/packages.config
